[BUGFIX] add pid correctly in lockfile

+++ ceof_runs.c
@@ -43,21 +43,27 @@ int ceof_runs()
    memset(pid_c, '\0', EOF_L_PIDLEN);
 
    if(fileexists(opt.pidfile)) {
+      printf("file already exists...\n");
       if(!openreadclosestatic(pid_c, opt.pidfile, EOF_L_PIDLEN)) {
          return -1;
       }
+      printf("saved pid is %s\n", pid_c);
       pid = strtoul(pid_c, NULL, 10);
-
-      if(kill(pid, 0) == 0)  {
-         return pid;
-      } else {
-         printf("Removing stale pidfile %s (%d).\n", opt.pidfile, pid);
-         if(unlink(opt.pidfile) == -1) {
-            return -1;
+      
+      if(pid) {
+         if(kill(pid, 0) == 0)  {
+            return pid;
          }
+      }
+      printf("Removing stale pidfile %s.\n", opt.pidfile);
+      if(unlink(opt.pidfile) == -1) {
+         return -1;
       }
    }
 
+   /* our pidfile! */
+   pid = getpid();
+   snprintf(pid_c, EOF_L_PIDLEN, "%ld", (long) pid);
    if(!openwriteclose(opt.pidfile, pid_c, 7)) {
       return -1;
    }
